+++ man/examples/example-plot_xyt.R
@@ -18,41 +18,41 @@ if (patter_run(.julia = FALSE, .geospatial = TRUE)) {
 
   #### Example (1): Plot selected samples
   # Plot particles from the forward filter
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = fwd,
              .steps = 1L)
   # Plot particles from the backward filter
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = bwd,
              .steps = 1L)
   # Plot smoothed particles
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = smo,
              .steps = 1L)
 
   #### Example (2): Plot multiple time steps
   # Specify selected steps
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = smo,
              .steps = 1:4L)
   # Plot all steps (default: .step = NULL)
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = fwd)
   # Use `.prompt = TRUE`
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = smo,
              .steps = 1:4L,
              .prompt = TRUE)
 
   #### Example (3): Customise the background map via `.add_surface`
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = smo,
              .steps = 1L,
              .add_surface = list(col = rev(terrain.colors(256L))))
 
   #### Example (4): Customise the points
   # Use `.add_points`
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = smo,
              .steps = 1L,
              .add_points = list(pch = ".", col = "red"))
@@ -63,26 +63,26 @@ if (patter_run(.julia = FALSE, .geospatial = TRUE)) {
   smo[, cell_id := terra::extract(map, cbind(x, y))]
   fwd[, col := ifelse(cell_id %in% smo$cell_id, "black", "red")]
   bwd[, col := ifelse(cell_id %in% smo$cell_id, "black", "red")]
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = rbind(fwd, bwd),
              .steps = 1L,
              .add_points = list(pch = "."))
 
   #### Example (5): Add additional map layers
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = smo,
              .steps = 1L,
              .add_layer = function(t) mtext(side = 4, "Depth (m)", line = -4))
 
   #### Example (6): Write images to file
   # Write images in serial
-  pf_plot_xy(.map = map,
+  plot_xyt(.map = map,
              .coord = smo,
              .steps = 1:4L,
              .png = list(filename = frames))
   # Use a fork cluster
   if (.Platform$OS.type == "unix") {
-    pf_plot_xy(.map = map,
+    plot_xyt(.map = map,
                .coord = smo,
                .steps = 1:4L,
                .png = list(filename = frames),