+++ tests/testthat/test-plots.R
@@ -1,4 +1,4 @@
-test_that("pf_plot_xy() works", {
+test_that("plot_xyt() works", {
 
   skip_if_not(patter_run(.julia = FALSE, .geospatial = TRUE))
 
@@ -18,12 +18,12 @@ test_that("pf_plot_xy() works", {
     arrange(path_id, timestep) |>
     mutate(cex = 1.5) |>
     as.data.table()
-  pf_plot_xy(.map = map,
-             .coord = fwd,
-             .steps = NULL,
-             .png = list(),
-             .add_points = list(col = "red"),
-             .add_layer = function(t) points(origin))
+  plot_xyt(.map = map,
+           .coord = fwd,
+           .steps = NULL,
+           .png = list(),
+           .add_points = list(col = "red"),
+           .add_layer = function(t) points(origin))
   setwd(wd)
   output   <- list.files(con)
   expected <- paste0(1:2, ".png")